pyproject.toml
[tool.poetry]
name = "flask-scaffold"
version = "0.1.0"
description = "A scaffold to speed up launching a flask project"
authors = [
    "Weiliang Li <to.be.impressive@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.7"

# flask
APIFlask = "^0.10.0"
flask = "^2.0.2"
flask-sqlalchemy = "^2.5.1"
flask-migrate = "^3.1.0"

# server
gunicorn = "^20.1.0"
aiohttp = "^3.8.1"
aiohttp_wsgi = "^0.9.1"

# db
psycopg2 = "^2.9.2"
redis = "^4.0.1"
hiredis = "^2.0.0"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
ipython = "^7.16.1"
black = "^20.8b1"
flake8 = "^4.0.1"
pre-commit = "^2.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
profile = "black"
